--- number_trivia_app/lib/features/number_trivia/presentation/bloc/bloc.dart
@@ -1,3 +1,0 @@
-export 'number_trivia_bloc.dart';
-export 'number_trivia_event.dart';
-export 'number_trivia_state.dart';
