--- number_trivia_app/lib/features/number_trivia/presentation/bloc/number_trivia_event.dart
@@ -1,21 +1,0 @@
-import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
-import 'package:flutter/material.dart';
-
-@immutable
-abstract class NumberTriviaEvent extends Equatable {
-
-  @override
-  List<Object> get props => [];
-}
-
-class GetTriviaForConcreteNumber extends NumberTriviaEvent {
-  final String numberString;
-
-  GetTriviaForConcreteNumber(this.numberString);
-
-  @override
-  List<Object> get props => [numberString];
-}
-
-class GetTriviaForRandomNumber extends NumberTriviaEvent {}
