--- number_trivia_app/lib/features/number_trivia/presentation/bloc/number_trivia_state.dart
@@ -1,31 +1,0 @@
-import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
-
-import '../../domain/entities/number_trivia.dart';
-
-abstract class NumberTriviaState extends Equatable {
-  @override
-  List<Object> get props => [];
-}
-
-class Empty extends NumberTriviaState {}
-
-class Loading extends NumberTriviaState {}
-
-class Loaded extends NumberTriviaState {
-  final NumberTrivia trivia;
-
-  Loaded({@required this.trivia});
-
-  @override
-  List<Object> get props => [trivia];
-}
-
-class Error extends NumberTriviaState {
-  final String message;
-
-  Error({@required this.message});
-
-  @override
-  List<Object> get props => [message];
-}
