+++ number_trivia_app/lib/features/number_trivia/presentation/states/number_trivia_error_state.dart
@@ -1,0 +1,11 @@
+import 'number_trivia_state.dart';
+import 'package:meta/meta.dart';
+
+class Error extends NumberTriviaState {
+  final String message;
+
+  Error({@required this.message});
+
+  @override
+  List<Object> get props => [message];
+}
