+++ number_trivia_app/lib/features/number_trivia/presentation/states/number_trivia_loaded_state.dart
@@ -1,0 +1,12 @@
+import '../../domain/entities/number_trivia.dart';
+import 'number_trivia_state.dart';
+import 'package:meta/meta.dart';
+
+class Loaded extends NumberTriviaState {
+  final NumberTrivia trivia;
+
+  Loaded({@required this.trivia});
+
+  @override
+  List<Object> get props => [trivia];
+}
