+++ number_trivia_app/lib/features/number_trivia/presentation/states/number_trivia_loading_state.dart
@@ -1,0 +1,3 @@
+import 'number_trivia_state.dart';
+
+class Loading extends NumberTriviaState {}
