+++ number_trivia_app/lib/features/number_trivia/presentation/states/number_trivia_state.dart
@@ -1,0 +1,8 @@
+import 'package:equatable/equatable.dart';
+
+abstract class NumberTriviaState extends Equatable {
+  @override
+  List<Object> get props => [];
+
+  get message => null;
+}
