+++ number_trivia_app/lib/features/number_trivia/presentation/widgets/widgets.dart
@@ -1,4 +1,3 @@
 export 'loading_widget.dart';
 export 'message_display.dart';
-export 'trivia_controls.dart';
 export 'trivia_display.dart';
